+++ include/compiler/output_x86_64.h
@@ -6,12 +6,12 @@
 void x86_64_Output(Operation operation);
 
 // c-stack
-void x86_64_StackPush(long num);
+void x86_64_StackPush(int num);
 
 // c-ctrl
-void x86_64_CtrlCall(long num);
+void x86_64_CtrlCall(int num);
 
 // c-tag
-void x86_64_TagDef(long num);
+void x86_64_TagDef(int num);
 
-void x86_64_TagJmp(long num);
+void x86_64_TagJmp(int num);
