+++ src/compiler/read_num.c
@@ -2,10 +2,10 @@
 
 #include <stdlib.h>
 
-bool ReadNum(char **program_ptr, void function(long num)){
+bool ReadNum(char **program_ptr, void function(int num)){
     char *program = *program_ptr;
     if ('0' <= *program && *program <= '9') {
-        long num = strtol(program, &program, 10);
+        int num = strtol(program, &program, 10);
         function(num);
         return true;
     } else {
